extract reserved children into component
