+++ components/list/ListItemMiddle.jsx
@@ -1,30 +1,14 @@
 import React from 'react';
 import style from './style';
-import ListItemCaption from './ListItemCaption'
-import ListItemLegend from './ListItemLegend'
+import ListItemCaption from './ListItemCaption';
+import ListItemLegend from './ListItemLegend';
+import ReservedChildren from '../reserved_children/ReservedChildren';
 
 class ListItemMiddle extends React.Component {
-  reservedChildrenTypes = new Set([ListItemCaption, ListItemLegend]);
-
-  isReservedChild (child) {
-    return child && this.reservedChildrenTypes.has(child.type);
-  }
-
-  reservedChildren () {
-    let children = {};
-    React.Children.forEach(this.props.children, (child) => {
-      if (this.isReservedChild(child)) {
-        children[child.type] = child;
-      }
-    });
-
-    return children;
-  }
-
   render () {
     const defaultCaption = <ListItemCaption> {this.props.caption} </ListItemCaption>;
     const defaultLegend = <ListItemLegend> {this.props.legend} </ListItemLegend>;
-    const reservedChildren = this.reservedChildren();
+    const reservedChildren = this.reservedChildrenByType();
 
     return (
       <span className={style.text}>
@@ -35,4 +19,4 @@ class ListItemMiddle extends React.Component {
   }
 }
 
-export default ListItemMiddle;
+export default ReservedChildren(ListItemMiddle, [ListItemCaption, ListItemLegend]);
